job_scraper/requirements.txt
Django==5.0.7 
requests==2.31.0 
beautifulsoup4==4.12.2 
google-generativeai==0.4.0 
